+++ NEWS.md
@@ -1,3 +1,6 @@
+# uncrtnty development version
+* Bug fix : `low_to_matrix` fill the lower matrix properly.
+
 # uncrtnty 0.1.0.9001
 
 * Added a `NEWS.md` file to track changes to the package.